+++ backend/canisters/local_group_index/api/src/updates/mod.rs
@@ -10,3 +10,4 @@ pub mod c2c_set_max_concurrent_group_upgrades;
 pub mod c2c_trigger_upgrade;
 pub mod c2c_upgrade_community_canister_wasm;
 pub mod c2c_upgrade_group_canister_wasm;
+pub mod push_events;