+++ backend/canisters/local_group_index/impl/src/lib.rs
@@ -1,9 +1,13 @@
 use crate::model::local_community_map::LocalCommunityMap;
 use candid::Principal;
 use canister_state_macros::canister_state;
+use event_sink_client::{EventSinkClient, EventSinkClientBuilder, EventSinkClientInfo};
+use event_sink_client_cdk_runtime::CdkRuntime;
+use event_sink_utils::EventDeduper;
 use model::local_group_map::LocalGroupMap;
 use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
+use std::time::Duration;
 use types::{
     BuildVersion, CanisterId, CanisterWasm, ChunkedCanisterWasm, Cycles, Milliseconds, TimestampMillis, Timestamped, UserId,
 };
@@ -11,6 +15,7 @@ use utils::canister;
 use utils::canister::{CanistersRequiringUpgrade, FailedUpgradeCount};
 use utils::consts::CYCLES_REQUIRED_FOR_UPGRADE;
 use utils::env::Environment;
+use utils::time::MINUTE_IN_MS;
 
 mod guards;
 mod lifecycle;
@@ -64,6 +69,8 @@ impl RuntimeState {
     pub fn metrics(&self) -> Metrics {
         let group_upgrades_metrics = self.data.groups_requiring_upgrade.metrics();
         let community_upgrades_metrics = self.data.communities_requiring_upgrade.metrics();
+        let event_sink_client_info = self.data.event_sink_client.info();
+        let event_relay_canister_id = event_sink_client_info.event_sink_canister_id;
 
         Metrics {
             memory_used: utils::memory::used(),
@@ -89,6 +96,7 @@ impl RuntimeState {
             group_upgrade_concurrency: self.data.group_upgrade_concurrency,
             max_concurrent_community_upgrades: self.data.max_concurrent_community_upgrades,
             community_upgrade_concurrency: self.data.community_upgrade_concurrency,
+            event_sink_client_info,
             canister_ids: CanisterIds {
                 user_index: self.data.user_index_canister_id,
                 group_index: self.data.group_index_canister_id,
@@ -97,6 +105,7 @@ impl RuntimeState {
                 proposals_bot: self.data.proposals_bot_user_id.into(),
                 escrow_canister_id: self.data.escrow_canister_id,
                 cycles_dispenser: self.data.cycles_dispenser_canister_id,
+                event_relay: event_relay_canister_id,
             },
         }
     }
@@ -127,7 +136,18 @@ struct Data {
     pub max_concurrent_community_upgrades: u32,
     pub community_upgrade_concurrency: u32,
     pub video_call_operators: Vec<Principal>,
+    #[serde(default = "event_sink_client")]
+    pub event_sink_client: EventSinkClient<CdkRuntime>,
+    #[serde(default)]
+    pub event_deduper: EventDeduper,
     pub rng_seed: [u8; 32],
+}
+
+fn event_sink_client() -> EventSinkClient<CdkRuntime> {
+    let event_relay_canister_id = CanisterId::from_text("6ofpc-2aaaa-aaaaf-biibq-cai").unwrap();
+    EventSinkClientBuilder::new(event_relay_canister_id, CdkRuntime::default())
+        .with_flush_delay(Duration::from_millis(MINUTE_IN_MS))
+        .build()
 }
 
 impl Data {
@@ -142,6 +162,7 @@ impl Data {
         cycles_dispenser_canister_id: CanisterId,
         proposals_bot_user_id: UserId,
         escrow_canister_id: CanisterId,
+        event_relay_canister_id: CanisterId,
         video_call_operators: Vec<Principal>,
         canister_pool_target_size: u16,
         test_mode: bool,
@@ -171,6 +192,10 @@ impl Data {
             community_upgrade_concurrency: 2,
             rng_seed: [0; 32],
             video_call_operators,
+            event_sink_client: EventSinkClientBuilder::new(event_relay_canister_id, CdkRuntime::default())
+                .with_flush_delay(Duration::from_millis(MINUTE_IN_MS))
+                .build(),
+            event_deduper: EventDeduper::default(),
         }
     }
 }
@@ -200,6 +225,7 @@ pub struct Metrics {
     pub group_upgrade_concurrency: u32,
     pub max_concurrent_community_upgrades: u32,
     pub community_upgrade_concurrency: u32,
+    pub event_sink_client_info: EventSinkClientInfo,
     pub canister_ids: CanisterIds,
 }
 
@@ -212,4 +238,5 @@ pub struct CanisterIds {
     pub proposals_bot: CanisterId,
     pub escrow_canister_id: CanisterId,
     pub cycles_dispenser: CanisterId,
+    pub event_relay: CanisterId,
 }