+++ backend/canisters/user/impl/src/updates/c2c_send_messages.rs
@@ -1,71 +1,15 @@
 use crate::updates::send_message::register_timer_jobs;
-use crate::{mutate_state, read_state, run_regular_jobs, RuntimeState};
-use canister_api_macros::update_msgpack;
+use crate::{mutate_state, read_state, RuntimeState};
 use canister_tracing_macros::trace;
 use chat_events::{MessageContentInternal, PushMessageArgs, Reader, ReplyContextInternal};
+use event_sink_client::EventBuilder;
 use ic_cdk_macros::update;
 use rand::Rng;
 use types::{
-    CanisterId, DirectMessageNotification, EventWrapper, Message, MessageId, MessageIndex, Notification, TimestampMillis,
-    UserId,
+    CanisterId, DirectMessageNotification, EventWrapper, Message, MessageEventPayload, MessageId, MessageIndex, Notification,
+    TimestampMillis, UserId,
 };
-use user_canister::c2c_send_messages_v2::{Response::*, *};
-
-#[update_msgpack]
-#[trace]
-async fn c2c_send_messages_v2(args: Args) -> Response {
-    c2c_send_messages_with_sender_check(args).await
-}
-
-async fn c2c_send_messages_with_sender_check(args: Args) -> Response {
-    run_regular_jobs();
-
-    let sender_user_id = match read_state(get_sender_status) {
-        SenderStatus::Ok(user_id) => user_id,
-        SenderStatus::Blocked => return Blocked,
-        SenderStatus::UnknownUser(local_user_index_canister_id, user_id) => {
-            if !verify_user(local_user_index_canister_id, user_id, false).await {
-                panic!("This request is not from an OpenChat user");
-            }
-            user_id
-        }
-    };
-
-    mutate_state(|state| c2c_send_messages_impl(args, sender_user_id, state))
-}
-pub(crate) fn c2c_send_messages_impl(args: Args, sender_user_id: UserId, state: &mut RuntimeState) -> Response {
-    let now = state.env.now();
-    for message in args.messages {
-        // Messages sent c2c can be retried so the same messageId may be received multiple
-        // times, so here we skip any messages whose messageId already exists.
-        if let Some(chat) = state.data.direct_chats.get(&sender_user_id.into()) {
-            if chat.events.contains_message_id(None, message.message_id) {
-                continue;
-            }
-        }
-
-        handle_message_impl(
-            sender_user_id,
-            HandleMessageArgs {
-                message_id: Some(message.message_id),
-                sender_message_index: Some(message.sender_message_index),
-                sender_name: args.sender_name.clone(),
-                sender_display_name: args.sender_display_name.clone(),
-                content: message.content,
-                replies_to: message.replies_to,
-                forwarding: message.forwarding,
-                correlation_id: message.correlation_id,
-                is_bot: false,
-                sender_avatar_id: args.sender_avatar_id,
-                now,
-            },
-            message.message_filter_failed.is_some(),
-            state,
-        );
-    }
-
-    Success
-}
+use user_canister::C2CReplyContext;
 
 #[update]
 #[trace]
@@ -104,12 +48,12 @@ async fn c2c_handle_bot_messages(
                     content: MessageContentInternal::from_initial(message.content, now).unwrap(),
                     replies_to: None,
                     forwarding: false,
-                    correlation_id: 0,
                     is_bot: true,
                     sender_avatar_id: None,
                     now,
                 },
                 false,
+                true,
                 state,
             );
         }
@@ -125,7 +69,6 @@ pub(crate) struct HandleMessageArgs {
     pub content: MessageContentInternal,
     pub replies_to: Option<C2CReplyContext>,
     pub forwarding: bool,
-    pub correlation_id: u64,
     pub is_bot: bool,
     pub sender_avatar_id: Option<u128>,
     pub now: TimestampMillis,
@@ -168,6 +111,7 @@ pub(crate) fn handle_message_impl(
     sender: UserId,
     args: HandleMessageArgs,
     mute_notification: bool,
+    push_message_sent_event: bool,
     state: &mut RuntimeState,
 ) -> EventWrapper<Message> {
     let replies_to = convert_reply_context(args.replies_to, sender, state);
@@ -181,7 +125,7 @@ pub(crate) fn handle_message_impl(
         mentioned: Vec::new(),
         replies_to,
         forwarded: args.forwarding,
-        correlation_id: args.correlation_id,
+        correlation_id: 0,
         now: args.now,
     };
 
@@ -215,6 +159,8 @@ pub(crate) fn handle_message_impl(
         if args.is_bot {
             chat.mark_read_up_to(message_event.event.message_index, false, args.now);
         }
+
+        let this_canister_id = state.env.canister_id();
         if !mute_notification && !chat.notifications_muted.value && !state.data.suspended.value {
             let content = &message_event.event.content;
             let notification = Notification::DirectMessage(DirectMessageNotification {
@@ -231,9 +177,22 @@ pub(crate) fn handle_message_impl(
                 crypto_transfer: content.notification_crypto_transfer_details(&[]),
             });
 
-            let recipient = state.env.canister_id().into();
+            let recipient = this_canister_id.into();
 
             state.push_notification(recipient, notification);
+        }
+
+        if push_message_sent_event {
+            state.data.event_sink_client.push(
+                EventBuilder::new("message_sent", args.now)
+                    .with_user(sender.to_string())
+                    .with_source(this_canister_id.to_string())
+                    .with_json_payload(&MessageEventPayload {
+                        message_type: message_event.event.content.message_type(),
+                        sender_is_bot: false,
+                    })
+                    .build(),
+            );
         }
     }