+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -389,6 +389,8 @@ export interface ProposalCommon {
     summary: string;
     proposer: string;
     payloadTextRendering?: string;
+    minYesPercentageOfTotal: number;
+    minYesPercentageOfExercised: number;
 }
 
 export type ManageNeuronResponse =