+++ backend/canisters/community/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Implement ability to push events from Community canisters ([#5436](https://github.com/open-chat-labs/open-chat/pull/5436))
+- Push event each time a message is sent ([#5439](https://github.com/open-chat-labs/open-chat/pull/5439))
+
 ### Changed
 
 - Use ICRC1 for ICP transactions between users ([#5426](https://github.com/open-chat-labs/open-chat/pull/5426))